Remove legacy session store
